# src/Liny.Core/Models/AttachModuleResponse.cs
/*
 * Merged API
 *
 * Merged OpenAPI specification
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using System.Text.Json;
using Liny.Core.Converters;

namespace Liny.Core.Models
{ 
    /// <summary>
    /// Attach by operation of the module channel provider
    /// </summary>
    [DataContract]
    public partial class AttachModuleResponse : IEquatable<AttachModuleResponse>
    {
        /// <summary>
        /// User ID of the bot on the LINE Official Account.
        /// </summary>
        /// <value>User ID of the bot on the LINE Official Account.</value>
        [Required]
        [DataMember(Name="bot_id", EmitDefaultValue=false)]
        public string BotId { get; set; }

        /// <summary>
        /// Permissions (scope) granted by the LINE Official Account admin.
        /// </summary>
        /// <value>Permissions (scope) granted by the LINE Official Account admin.</value>
        [Required]
        [DataMember(Name="scopes", EmitDefaultValue=false)]
        public List<string> Scopes { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class AttachModuleResponse {\n");
            sb.Append("  BotId: ").Append(BotId).Append("\n");
            sb.Append("  Scopes: ").Append(Scopes).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            var options = new JsonSerializerOptions
            {
                WriteIndented = true
            };

            return JsonSerializer.Serialize(this, options);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (obj is null) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((AttachModuleResponse)obj);
        }

        /// <summary>
        /// Returns true if AttachModuleResponse instances are equal
        /// </summary>
        /// <param name="other">Instance of AttachModuleResponse to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(AttachModuleResponse other)
        {
            if (other is null) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    BotId == other.BotId ||
                    BotId != null &&
                    BotId.Equals(other.BotId)
                ) && 
                (
                    Scopes == other.Scopes ||
                    Scopes != null &&
                    other.Scopes != null &&
                    Scopes.SequenceEqual(other.Scopes)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (BotId != null)
                    hashCode = hashCode * 59 + BotId.GetHashCode();
                    if (Scopes != null)
                    hashCode = hashCode * 59 + Scopes.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(AttachModuleResponse left, AttachModuleResponse right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(AttachModuleResponse left, AttachModuleResponse right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}